fix: add more precision to draw_line

+++ src/draw_line.c
@@ -6,7 +6,7 @@
 /*   By: troberts <troberts@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/04 19:40:16 by troberts          #+#    #+#             */
-/*   Updated: 2023/01/05 00:29:23 by troberts         ###   ########.fr       */
+/*   Updated: 2023/01/11 01:06:17 by troberts         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,6 +21,11 @@ void	drawline(t_img img, t_map_point start, t_map_point end)
 	int error = dx + dy;
 	t_map_point	point;
 
+	// if (start.x < 0 && start.y < 0 && end.x < 0 && end.y < 0)
+	// 	return ;
+	// if (start.x > WIN_W - 1 && start.y > WIN_H - 1 && \
+	// 	end.x > WIN_W - 1 && end.y > WIN_H - 1) // A confirmer formule
+	// 	return ;
 	point = start;
 	while (true)
 	{
@@ -32,14 +37,14 @@ void	drawline(t_img img, t_map_point start, t_map_point end)
 			if ((int)point.x == (int)end.x)
 				break ;
 			error = error + dy;
-			point.x = (int)point.x + sx;
+			point.x = point.x + sx;
 		}
 		if ((error * 2) <= dx)
 		{
 			if ((int)point.y == (int)end.y)
 				break ;
 			error = error + dx;
-			point.y = (int)point.y + sy;
+			point.y = point.y + sy;
 		}
 	}
 }
